ibverbs: Fix cli device matching
Previously the device matching in the ibverbs module selected the
first available device it came across, whether or not it matched the
device or port specified on the command line.

This change will correctly match the device and the port if they are
specified, or use the first available if not.